requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyMuPDF
pymupdf4llm
ollama
nltk